# services/money/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sagolik Close money service (internal)
  version: 0.1.0
  description: |
    Internal API reachable only over mutual TLS from allowed workloads
    (e.g. spiffe://sagolik/web). Every /v1 call carries a user assertion:
    an Ed25519 compact JWS (alg EdDSA), at most 60 seconds valid, single use.
    Sagolik Close never holds client funds; balances mirror what the licensed
    escrow partner reports.
servers:
  - url: https://money.internal:8443
security:
  - userAssertion: []
paths:
  /v1/session:
    get:
      summary: Verify the caller's assertion and describe the principal
      responses:
        "200":
          description: Verified caller
          content:
            application/json:
              schema:
                type: object
                required: [userId, aal]
                properties:
                  userId: { type: string, format: uuid }
                  aal: { type: string, enum: [aal1, aal2] }
                  transactionId: { type: string, format: uuid }
                  role: { type: string }
                  stepUpAgeSeconds: { type: integer, minimum: 0 }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /v1/transactions/{id}/funds:
    get:
      summary: Funds summary for a transaction (audited)
      description: The assertion must be scoped to this transaction and carry a role allowed to view money.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Balances per currency, in cents
          content:
            application/json:
              schema:
                type: object
                required: [transactionId, balances]
                properties:
                  transactionId: { type: string, format: uuid }
                  balances:
                    type: array
                    items: { $ref: "#/components/schemas/Balance" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/transactions/{id}/instructions:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Payment instruction versions (masked), newest first per purpose
      responses:
        "200":
          description: Versions with derived status
          content:
            application/json:
              schema: { type: object, required: [instructions], properties: { instructions: { type: array, items: { $ref: "#/components/schemas/Instruction" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: Record a new version (escrow officer; AAL2 + step-up within 5 minutes)
      description: A changed version is high risk (critical within 72 hours of closing) and gets a cooling-off period before it can be used.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [purpose, beneficiaryName, bankName, routingNumber, accountNumber, currency]
              additionalProperties: false
              properties:
                purpose: { type: string, enum: [earnest_money_to_escrow, closing_funds_to_escrow, seller_proceeds, loan_payoff] }
                beneficiaryName: { type: string, minLength: 2, maxLength: 140 }
                bankName: { type: string, minLength: 2, maxLength: 140 }
                routingNumber: { type: string, pattern: "^[0-9]{9}$", description: US ABA routing number (checksum validated) }
                accountNumber: { type: string, pattern: "^[0-9 ]{4,20}$" }
                currency: { type: string, pattern: "^[A-Z]{3}$" }
                hoursToClosing: { type: number }
      responses:
        "201": { description: Created (masked), content: { application/json: { schema: { type: object, properties: { instruction: { $ref: "#/components/schemas/Instruction" } } } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/StepUpRequired" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/instructions/{iid}/verify:
    post:
      summary: Second-person, out-of-band verification (title/escrow officer; not the author; step-up)
      parameters:
        - { name: iid, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method]
              additionalProperties: false
              properties:
                method: { type: string, enum: [out_of_band_call, in_person, provider_attested] }
                reference: { type: string, maxLength: 200 }
      responses:
        "200": { description: Verified, content: { application/json: { schema: { type: object, properties: { instruction: { $ref: "#/components/schemas/Instruction" } } } } } }
        "403": { description: Step-up required, or the caller authored these instructions, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Superseded or already decided, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /v1/instructions/{iid}/reveal:
    post:
      summary: Full wire details for the payer (buyer; verified, past cooling-off, step-up; audited)
      parameters:
        - { name: iid, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The only response that carries a full account number
          content:
            application/json:
              schema: { type: object, required: [instruction, accountNumber], properties: { instruction: { $ref: "#/components/schemas/Instruction" }, accountNumber: { type: string } } }
        "403": { description: Step-up required, or inside the cooling-off period, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Not verified or superseded, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /v1/transactions/{id}/ledger/movements:
    post:
      summary: Record what the escrow system shows (escrow officer; step-up; idempotent)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, amount, currency, reference, idempotencyKey]
              additionalProperties: false
              properties:
                kind: { type: string, enum: [expectation, receipt, disbursement] }
                amount: { type: integer, minimum: 1, description: Cents }
                currency: { type: string, pattern: "^[A-Z]{3}$" }
                reference: { type: string, minLength: 3, maxLength: 120 }
                idempotencyKey: { type: string, minLength: 8, maxLength: 200 }
      responses:
        "201": { description: Recorded; returns derived balances }
        "200": { description: Same idempotency key and contents — nothing new recorded }
        "403": { $ref: "#/components/responses/StepUpRequired" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Key reused with different contents, or payout exceeds funds reported held, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
components:
  securitySchemes:
    userAssertion:
      type: http
      scheme: bearer
      bearerFormat: JWS (EdDSA, 60s, single use)
  schemas:
    Balance:
      type: object
      required: [currency, expected, outstanding, received, disbursed]
      properties:
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        expected: { type: integer, description: Total the buyer has been asked to send to escrow }
        outstanding: { type: integer, description: Still due; negative means more than expected was received }
        received: { type: integer, description: Currently held by escrow, as reported }
        disbursed: { type: integer }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
    Instruction:
      type: object
      properties:
        id: { type: string, format: uuid }
        transactionId: { type: string, format: uuid }
        purpose: { type: string }
        version: { type: integer }
        previousId: { type: [string, "null"], format: uuid }
        beneficiaryName: { type: string }
        bankName: { type: string }
        routingNumber: { type: string }
        accountMask: { type: string, pattern: "^[0-9]{4}$" }
        currency: { type: string }
        riskLevel: { type: string, enum: [low, medium, high, critical] }
        effectiveAfter: { type: [string, "null"], format: date-time }
        status: { type: string, enum: [pending_verification, verified, rejected, superseded] }
        usable: { type: boolean }
        verifiedBy: { type: [string, "null"], format: uuid }
        verifiedAt: { type: [string, "null"], format: date-time }
        verificationMethod: { type: [string, "null"] }
  responses:
    BadRequest:
      description: Invalid input
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    StepUpRequired:
      description: The caller must confirm it's them (AAL2 + step-up within 5 minutes)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthenticated:
      description: Missing, invalid, expired or replayed assertion
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Unknown transaction or no access (indistinguishable by design)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }